// OrderApp.Client/src/app/core/models/DTOs/IUserByCreateDTO.ts
export interface IUserByCreateDTO {
    citizenId: number
    fullName: string
    email: string
    mobilePhones: string
    status: boolean
    birthDate: string
    gender: number
    recordDate: string
    address: string
    notes: string
    updateContactDate: string
    password: string
    imagePath: string
  }
  